rename: 기존 close버튼 파일명을 색상 표시해서 수정
